docker/frontend/requirements.txt
python-dotenv
streamlit
requests
streamlit-option-menu
streamlit
streamlit_extras